NPC chase/aggro now implemented

+++ Main/src/com/grenadelawnchair/games/tbb/model/GameCharacter.java
@@ -146,7 +146,7 @@ public class GameCharacter {
 		return name;
 	}
 	
-	public double getHealth(){
+	public int getHealth(){
 		return health;
 	}
